kernel:hmq: user can enable/disable interrupt mode for hmq_in
For the hmq_in (from host to core) the interrupts are not really
helpful and sometimes it can be a problem. With this patch the
user can enable/disable the interrupt mode according to his
preference. By default it's disable.

Without interrupt we do not know when the core will consume
the message, so the driver will wait a bit (programmable) to
give time to the core to consume the message.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>