wrtd:lib|rt: re-write function to get trigger
The behavior was not clear because it was overwriting the same values.

From the real-time application we were sending both a trigger and,
if any, its condition (which is a trigger itself). On library side
we were reading both: the condition (if any) and than the trigger.

But when the user was interested in just the trigger we were
overwriting the condition.

So with this patch we send (RT) and receive (lib) only what we requested

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>