wrtd:[incompat] remove hash table from RT application
    "premature optimization is the root of all evil"

The hash implementation was buggy. Difficult to understand because it
includes also a memory allocator.

My humble opinion is that we do not need such optimization for a very
limited entries set (128). I replaced the hash table with an ordered by
trigger-id array. Ordered because with an optimization you can use
a binary search to retrieve in few step the correct trigger.

NOTE: a draft of the binary search algorithm is there but not tested,
that's way is inactive code.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>