wrtd:lib: make time values consistent within the library
The output dead-time was using nano-seconds, while all the other
functions in the library use pico-seconds. There reason was that
for the output dead-time we have 8ns granularity so having a
pico-seconds value is not really usefull.

We decided to move to pico-seconds for to be consistent with the
library because this difference caused wrong assumptions on
application side, even on my own test tool XD

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>