kernel: [bugfix] concurrency problem in sync messages
The problem
It may happen to receive (interrupt) the synchronous answer before
setting the variable waiting_seq. The consequence is that the answer
is not detected.

The solution
The solution is to set this variable *before* sending the message

Note.
I noticed also that the spinlock is not necessary. this variable is written only
in this place and elsewhere it is only read.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>
Signed-off-by: Michel Arruat <michel.arruat@cern.ch>